required changes made
