+++ frontend/src/pages/Admin/Components/AddTeamMember/add-team-member.module.scss
@@ -134,6 +134,7 @@
   color: var(--secondary-color);
   font-weight: 700;
   font-size: medium;
+  color: gray;
 }
 
 .inside-add-team-member {
