+++ frontend/src/pages/Admin/Components/MangeTeams/ManageTeams.jsx
@@ -1,29 +1,8 @@
 import React from "react";
 import style from "./manage-teams.module.scss";
+import cards from "../../../../test_data/teams"
 
-export function ManageTeams(props) {
-  const cards = [
-    {
-      name: "Team 1",
-      icon: <i className='fas fa-users fa-lg'></i>,
-      description:"Elite",
-    },
-    {
-      name: "Team 2",
-      icon: <i class="fas fa-users fa-lg"></i>,
-      description:"Semi Elite",
-    },
-    {
-      name: "Team 3",
-      icon: <i class="fas fa-users fa-lg"></i>,
-      description:"Experts",
-    },
-    {
-      name: "Team 4",
-      icon: <i class="fas fa-users fa-lg"></i>,
-      description:"Operational",
-    },
-  ];
+export function ManageTeams() {
   return (
     <div>
       <h1 className={style["head"]}>Teams</h1>
